+++ internal/utils.go
@@ -3,6 +3,7 @@ package internal
 import (
 	"fmt"
 	"os"
+	"reflect"
 	"regexp"
 	"strconv"
 	"strings"
@@ -120,6 +121,56 @@ func isASCIIDigit(c byte) bool {
 
 func isASCIIUpper(c byte) bool {
 	return 'A' <= c && c <= 'Z'
+}
+
+func ResolveFeatureName(field reflect.StructField) (string, error) {
+	if tag := field.Tag.Get(NameTag); tag != "" {
+		return tag, nil
+	}
+	versioned := field.Tag.Get("versioned")
+	fieldName := ChalkpySnakeCase(field.Name)
+	if versioned == "true" {
+		parts := strings.Split(fieldName, "_")
+		nameErr := fmt.Errorf(
+			"versioned feature must have a version suffix `VN` at the"+
+				" end of the attribute name, but found '%s' instead",
+			fieldName,
+		)
+		if len(parts) == 1 {
+			return "", nameErr
+		}
+		lastPart := parts[len(parts)-1]
+		if !strings.HasPrefix(lastPart, "v") {
+			return "", nameErr
+		}
+		version := lastPart[1:]
+		prefix := strings.Join(parts[:len(parts)-1], "_")
+		if version == "1" {
+			fieldName = prefix
+		} else {
+			fieldName = prefix + "@" + version
+		}
+	} else if strings.HasPrefix(versioned, "default(") && strings.HasSuffix(versioned, ")") {
+		version := versioned[len("default(") : len(versioned)-len(")")]
+		_, convertErr := strconv.Atoi(version)
+		if convertErr != nil {
+			return "", fmt.Errorf(
+				"expected struct tag `versioned:\"default(N)\"` "+
+					"where N is an integer, but found %s instead",
+				versioned,
+			)
+		}
+		if version != "1" {
+			fieldName = fieldName + "@" + version
+		}
+	} else if versioned != "" {
+		return "", fmt.Errorf(
+			"expected struct tag `versioned:\"true\"` or `versioned:\"default(N)\"` "+
+				"where N is an integer, but found '%s' instead",
+			versioned,
+		)
+	}
+	return fieldName, nil
 }
 
 func Ptr[T any](value T) *T {
